Rename existing logger class
