+++ studio/app/common/core/snakemake/snakemake_executor.py
@@ -4,15 +4,15 @@ from typing import Dict
 
 from snakemake import snakemake
 
-from studio.app.common.core.logger import Logger
 from studio.app.common.core.snakemake.smk import SmkParam
+from studio.app.common.core.snakemake.smk_status_logger import SmkStatusLogger
 from studio.app.common.core.utils.filepath_creater import get_pickle_file, join_filepath
 from studio.app.common.core.workflow.workflow import Edge, Node
 from studio.app.dir_path import DIRPATH
 
 
 def snakemake_execute(workspace_id: str, unique_id: str, params: SmkParam):
-    logger = Logger(workspace_id, unique_id)
+    logger = SmkStatusLogger(workspace_id, unique_id)
     snakemake(
         DIRPATH.SNAKEMAKE_FILEPATH,
         forceall=params.forceall,
@@ -66,7 +66,7 @@ def delete_dependencies(
                 ),
             ]
         )
-        # print(pickle_filepath)
+        # logger.debug(pickle_filepath)
 
         if os.path.exists(pickle_filepath):
             os.remove(pickle_filepath)
